# packages/contracts/openapi/agent-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Course Tutor Agent API",
    "version": "1.0.0",
    "description": "Canonical v1 contract. x-implementation-status distinguishes implemented, partial and planned operations."
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/v1/courses/{course_id}": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}],
      "get": {
        "operationId": "getCourse",
        "x-implementation-status": "implemented",
        "responses": {
          "200": {"description": "Authorized course", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Course"}}}},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/v1/courses/{course_id}/chat": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}],
      "post": {
        "operationId": "streamCourseChat",
        "x-implementation-status": "partial",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}},
        "responses": {
          "200": {"description": "SSE events: token, citation, abstained, warning, done or error", "content": {"text/event-stream": {"schema": {"type": "string"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/v1/admin/courses": {
      "post": {
        "operationId": "createCourse",
        "x-implementation-status": "implemented",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCourseRequest"}}}},
        "responses": {"201": {"description": "Course and course run registered"}, "403": {"$ref": "#/components/responses/Forbidden"}}
      }
    },
...
    "/v1/courses/{course_id}/sources/{source_id}": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}, {"$ref": "#/components/parameters/SourceId"}],
      "get": {
        "operationId": "getCourseSource",
        "x-implementation-status": "implemented",
        "responses": {"200": {"description": "Authorized source metadata and anchors"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "CourseId": {"name": "course_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
      "SourceId": {"name": "source_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
    "responses": {
      "Unauthorized": {"description": "Authentication required"},
      "Forbidden": {"description": "Insufficient role, membership or tenant scope"},
      "NotFound": {"description": "Resource not found in the authorized scope"}
    },
    "schemas": {
...
      "Course": {"type": "object", "required": ["id", "tenant_id", "code", "name", "level"], "properties": {"id": {"type": "string", "format": "uuid"}, "tenant_id": {"type": "string", "format": "uuid"}, "programme_id": {"type": ["string", "null"], "format": "uuid"}, "course_run_id": {"type": ["string", "null"], "format": "uuid"}, "code": {"type": "string"}, "name": {"type": "string"}, "level": {"type": "string"}, "active_content_version_id": {"type": ["string", "null"], "format": "uuid"}}},
      "ChatRequest": {"type": "object", "additionalProperties": false, "required": ["query"], "properties": {"query": {"type": "string", "minLength": 1, "maxLength": 2000}, "session_id": {"type": ["string", "null"], "format": "uuid"}}},
...
      "CreateCourseRequest": {"type": "object", "additionalProperties": false, "required": ["programme_id", "code", "name", "level", "run_key", "source_path"], "properties": {"programme_id": {"type": "string", "format": "uuid"}, "code": {"type": "string"}, "name": {"type": "string"}, "level": {"enum": ["primary", "middle_school", "high_school", "undergraduate", "postgraduate"]}, "run_key": {"type": "string"}, "source_path": {"type": "string"}, "scan_interval_seconds": {"type": "integer", "minimum": 30, "default": 900}}},
...
    }
  }
}